requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.13.0
opencv-python-headless==4.8.1.78
matplotlib==3.7.2
numpy==1.24.3
pillow==10.0.0